% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2tfOptimizedDiffusion.R
\name{diffusionMap2}
\alias{diffusionMap2}
\title{Compute Diffusion Map for Signal Propagation}
\usage{
diffusionMap2(
  receptors,
  TFs,
  feedbackNodes,
  M,
  network,
  inputSignal = 0.99,
  nTicks = 2000,
  x = rep(1, nrow(network))
)
}
\arguments{
\item{receptors}{Character vector specifying receptor nodes.}

\item{TFs}{Character vector specifying transcription factor nodes.}

\item{feedbackNodes}{Character vector specifying nodes that act as signal sinks.}

\item{M}{A numeric matrix of gene expression data (rows = genes, cols = samples).}

\item{network}{A data frame representing the network topology with columns from and to.}

\item{inputSignal}{Numeric, amount of initial signal (default 0.99).}

\item{nTicks}{Integer, number of time steps for signal propagation (default 2000).}

\item{x}{Numeric vector, weights for network edges.}
}
\value{
A list containing:
\itemize{
\item \code{diffCurves}: Signal propagation curves for each sample.
\item \code{rowAnnotation}: Metadata linking curves to samples, TFs, and receptors.
}
}
\description{
Compute Diffusion Map for Signal Propagation
}
