% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2tfOptimizedDiffusion.R
\name{initializeSignallingNetwork}
\alias{initializeSignallingNetwork}
\title{Initialize a Signaling Network}
\usage{
initializeSignallingNetwork(network, nodeW)
}
\arguments{
\item{network}{A dataframe with columns \code{from} and \code{to}, specifying network edges.}

\item{nodeW}{A named numeric vector of gene expression levels for each node.}
}
\value{
A list containing:
\itemize{
\item \code{network}: A \code{data.table} object with edges defined by columns from and to.
\item \code{signal}: A named numeric vector typically initialized to zero. except for receptor
}
}
\description{
This function initializes a signaling network by assigning edge weights based on the product of connected node weights.
}
\examples{
network <- data.frame(from = c('A', 'B'), to = c('B', 'C'))
nodeW <- c(A = 1, B = 2, C = 3)
initializeSignallingNetwork(network, nodeW)
}
