% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r2tfData}
\alias{r2tfData}
\title{Example Data for r2tfDiffusion Package}
\format{
A list with the following elements:
\describe{
\item{network}{A dataframe representing the TLR4 signaling network (edges).}
\item{feedbackNodes}{A character vector of feedback or negative regulator genes.}
\item{tfAct}{Transcription factor activity changes ex vivo.}
\item{Mexvivo}{Ex vivo gene expression before stimulation.}
\item{Mclinical}{Gene expression in clinical samples before treatment.}
\item{patientData}{Data frame of clinical patient outcomes}
}
}
\usage{
data(r2tfData)
}
\description{
A list containing example network gene expression data and transcription factor activity andclinical outcomes
}
\examples{
data(r2tfData)
names(r2tfData)

}
\keyword{datasets}
