% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2tfOptimizedDiffusion.R
\name{signalOnFeedbackNetwork}
\alias{signalOnFeedbackNetwork}
\title{Simulate Signal Flow in a Feedback Network}
\usage{
signalOnFeedbackNetwork(
  network,
  outputNode,
  inputNode,
  feedbackNodes = NULL,
  inputSignal = 0.99,
  n = 2000
)
}
\arguments{
\item{network}{A list output from \code{initializeSignallingNetwork()}.}

\item{outputNode}{A character vector specifying the output nodes (TFs).}

\item{inputNode}{A character vector specifying the input nodes (e.g., receptors).}

\item{feedbackNodes}{A character vector specifying nodes that act as negative regulators.}

\item{inputSignal}{Numeric, amount of initial signal (default 0.99).}

\item{n}{Integer, number of iterations for signal diffusion (default 2000).}
}
\value{
A list containing:
\itemize{
\item \code{Enode}: A matrix tracking signal changes over time at output nodes.
\item \code{signal}: The final state of all nodes after diffusion.
}
}
\description{
This function simulates the propagation of a signal through a network with feedback regulation.
}
